Allow passing of cucumber command as ENV var
This will allow us to pass in a command like:

[[ $CUKES_OFF -eq 0 ]] && bundle exec cucumber

We can then control whether we run cukes via our commits as a commit directive.
When we add [cukes off] to a commit we can add a setup command on Semaphore that
parses the last git log entry and sets CUKES_OFF=1 if the directive is present.