Add the ability for dirs to be excluded from Rspec
Currently, 'spec/**/*_spec.rb' is the hardcoded path from which the
boosters gather and run specs. Agworld has the need to exclude a
directory from this path, in order to further split up how our tests are
run on Semaphore.
The environment variable BOOSTERS_EXCLUDE_PATH can now be set to exclude
a directory or file from running on the Rspec boosters.