Removed flaky command selection.
Previously the file distributor used regex to select commands from the
git log. However this would not work very well in practice as it would
get confused by multiple commands and potentially not react. This update
allows for it to respond to any number of commands, and is much more
stable.