Add ability to filter paths from manual load.
This update brings in manual balancing for the slowest specs in our
repo. The hope is that if the most time-intensive specs are properly
distributed, then semaphore can only screw up the rest so much.

Normally this manual file load would over-write any commit directives as
it is called after Distributor, but I added in a filter so the
directives take precedence.